+++ saleor/graphql/attribute/mutations/__init__.py
@@ -1,3 +1,4 @@
+from .attribute_bulk_create import AttributeBulkCreate
 from .attribute_create import AttributeCreate
 from .attribute_delete import AttributeDelete
 from .attribute_reorder_values import AttributeReorderValues
@@ -13,6 +14,7 @@ from .base_reorder_attributes import (
 __all__ = [
     "AttributeCreate",
     "AttributeDelete",
+    "AttributeBulkCreate",
     "AttributeReorderValues",
     "AttributeUpdate",
     "AttributeValueCreate",